informatino about statuses of bookings availbale across the whole client app
